main_test.go
package main

import (
	"bytes"
	"encoding/json"
	"os"
	"testing"

	sariftemplate "google3/experimental/CertifiedOSS/JsonToSarif/template" // Replace with the correct import path
)

func TestMain(t *testing.T) {
	// Test Data Setup
	testViolations := struct {
		Response struct {
			IACValidationReport struct {
				Violations []sariftemplate.Violation `json:"violations"`
			} `json:"iacValidationReport"`
		} `json:"response"`
	}{
		// ... (Same test data as before)
	}

	// Create a buffer to capture the output
	var outputBuffer bytes.Buffer

	// Redirect standard output to the buffer
	originalStdout := os.Stdout
	os.Stdout = &outputBuffer

	// Execute the main function (replace with the actual function call)
	// Assuming there's a function like `convertToSarif` that returns the SARIF output as []byte
	sarifData := convertToSarif(testViolations) // Replace with actual function call

	// Restore standard output
	os.Stdout = originalStdout

	// Parse the SARIF output from the buffer
	var sarifOutput sariftemplate.SarifOutput
	if err := json.Unmarshal(sarifData, &sarifOutput); err != nil {
		t.Fatalf("Error decoding SARIF output: %v", err)
	}

	// Assertions (add more specific assertions based on expected output)
	if len(sarifOutput.Runs) != 1 {
		t.Errorf("Expected 1 run, got %d", len(sarifOutput.Runs))
	}
	// ... Add more specific assertions for results, rules, etc.

	// You can also inspect the contents of outputBuffer if needed
	// t.Logf("Output Buffer Contents: %s", outputBuffer.String())
}

// (Add the convertToSarif function here if it's not already present)
func convertToSarif(violations interface{}) []byte {
	// ... (Your conversion logic here)

	// Marshal the SARIF output
	sarifJSON, err := json.MarshalIndent(sarifOutput, "", "  ")
	if err != nil {
		// Handle the error appropriately
	}

	return sarifJSON
}
